Code/pages/requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.0
black==21.9b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
cargo==0.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
entrypoints==0.3
eplot==0.1.2
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.24
greenlet==1.1.2
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
plotly==5.3.1
pluggy==1.0.0
prettytable==2.2.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psycopg2-binary==2.9.1
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
pyecharts==1.9.0
Pygments==2.10.0
pyodbc==4.0.32
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.3
pyxlsb==1.0.8
pyzmq==22.3.0
regex==2021.8.28
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.17.5
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.26
sqlalchemy-redshift==0.8.8
streamlit==1.14.0
streamlit-echarts==0.3.0
tdqm==0.0.1
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.8.0
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
XlsxWriter==3.0.1

folium==0.13.0
streamlit_folium==0.6.15
streamlit_tree_select==0.0.5
streamlit_extras==0.2.0
